+++ renderer/src/draw_commands/mod.rs
@@ -2,12 +2,10 @@ pub mod mesh2d_draw_command;
 pub mod mesh3d_draw_command;
 pub mod text_draw_command;
 
-use crate::mesh_layers::layers::Layers;
-use wgpu_canvas::render_modifier::SpatialData;
-use lyon::lyon_tessellation::LineJoin;
-use lyon::path::LineCap;
 use crate::buffer_pool::BufferPool;
 use crate::global_context::GlobalContext;
+use crate::mesh_layers::layers::Layers;
+use wgpu_canvas::render_modifier::SpatialData;
 
 pub(crate) struct DrawCommands {
     key: String,
